Finished adding the name for the job order endpoint and also did some renaming to invoices in the frontend
